refactor: put go code in src/ directory
