+++ .goreleaser.yml
@@ -21,10 +21,10 @@ builds:
       # - 386
     ldflags:
       ## auto inject commit and version at build time
-      - -X github.com/vdbulcke/oidc-client-demo/cmd.GitCommit={{ .ShortCommit }} 
-      - -X github.com/vdbulcke/oidc-client-demo/cmd.Version={{ .Version }}  
-      - -X github.com/vdbulcke/oidc-client-demo/cmd.Date={{ .CommitDate }}
-      - -X github.com/vdbulcke/oidc-client-demo/cmd.BuiltBy=goreleaser
+      - -X github.com/vdbulcke/oidc-client-demo/src/cmd.GitCommit={{ .ShortCommit }} 
+      - -X github.com/vdbulcke/oidc-client-demo/src/cmd.Version={{ .Version }}  
+      - -X github.com/vdbulcke/oidc-client-demo/src/cmd.Date={{ .CommitDate }}
+      - -X github.com/vdbulcke/oidc-client-demo/src/cmd.BuiltBy=goreleaser
 archives:
   - 
     ## Deprecated "replacements"
