+++ src/client/config_test.go
