+++ src/cmd/introspect.go
@@ -4,7 +4,7 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
-	oidcclient "github.com/vdbulcke/oidc-client-demo/oidc-client"
+	oidcclient "github.com/vdbulcke/oidc-client-demo/src/client"
 )
 
 // args var
